requirements.txt
requests
openai
python-dotenv
colorama
pywin32
mitreattack-python